23/04/24 Commit #3
Some more UI changes. I added a proper loading screen. You can now also
view the stations and tracks in the new menu at the top.

I still need to fix A LOT of bugs. There's an issue with the GUI when
you try to add and remove stops. For some reason the buttons just shift
down so I still have to figure that out.