Commit #5 30/04/24
Today I added occlusion culling to the overlays. I also improved the
rendering performance of the tracks by making the program only render
the longer tracks.
I also added the ability to open the html file from the program without
it opening automatically.
Finally, I added some messages which warn the user when they enter
invalid locations.
In the future I'd like to make the user be able to download station data
with the map.