Commit #6 02/05/24
I'm very happy with the progress.
The user now has the ability to download track and station data
themselves. I also completely changed the way files are stored. I now
let java serialize my objects and stores them in the files. This is more
convenient. I did have to replace every WayNode with a Long indicating
to the WayNode id. But it was worth it and I'm pretty sure this makes
the program use less memory too.
There's also a new loading screen with messages.

In the future I really want the user to be able to download data from
the map without needing to use the current GUI.