Commit #8 14/05/24
I patched a bug which caused the tracks to not load in and I patched a
bug which caused the program to crash when the user entered a number in
the input fields. The user can now input an id from the Overpass API
(way Id) and it will pathfind from that path.