Commit #9 22/09/24
I'd like to continue working on this project but I'll have to see if I
have the motivation to continue. For now I added the option to see the
tracks where the A* algorithm is searching through. I also finally
remembered to run the map on a separate thread :P
There's also some minor UI improvements.